Expand access to POSIX shared memory for PostgreSQL container
By default, the container runtime engine uses 64MiB of POSIX shared
memory, regardless of the resources that are given to the container.
Give this is inadequately low for larger PostgreSQL installs, its
necessary to expand the shared memory that the PostgreSQL container
has access to.

Using guidance from the OpenShift documentation[1], this patch
mounts the shared memory directory to the PostgreSQL container
to expand the access to POSIX shared memory.

Issue: #1358

[1] https://access.redhat.com/documentation/en-us/openshift_container_platform/3.11/html/developer_guide/dev-guide-shared-memory